exp1_linear_probe/umd_linear_probing/requirements.txt
torch>=2.1
torchvision>=0.16
scipy>=1.10
numpy>=1.23
matplotlib>=3.7
pyyaml>=6.0
open_clip_torch>=2.28
transformers>=4.40
safetensors>=0.4
einops>=0.7
sentencepiece>=0.1.99
diffusers>=0.30
accelerate>=0.24
segment-anything>=1.0
